+++ NEWS.md
@@ -1,3 +1,5 @@
+# workflows (development version)
+
 # workflows 0.2.4
 
 * `add_model()` and `update_model()` now use `...` to separate the required